+++ Changes.md
@@ -14,6 +14,7 @@ Improvements
 Fixes
 -----
 
+- PointInstancer : Fixed rendering of points and/or prototypes with zero scale.
 - Cycles : Fixed rendering of shaders written with `IECOREUSD_WRITE_CONFORMANT_OSL_SHADERS=1`.
 - NodeEditor : Fixed flicker when changing the node being viewed.
 
@@ -21,6 +22,11 @@ API
 ---
 
 - Switch, NameSwitch : Added support for subclassing in Python.
+
+API
+---
+
+- TestCase : Added Quatf support to `assertEqualWithAbsError()`.
 
 1.7.1.0 (relative to 1.7.0.0)
 =======